+++ app/src/androidTest/java/org/simple/clinic/di/TestAppComponent.kt
@@ -46,6 +46,7 @@ import org.simple.clinic.rules.ServerRegistrationAtFacilityRule
 import org.simple.clinic.security.pin.BruteForceProtectionAndroidTest
 import org.simple.clinic.signature.SignatureRepositoryAndroidTest
 import org.simple.clinic.storage.DaoWithUpsertAndroidTest
+import org.simple.clinic.storage.DatabaseEncryptorTest
 import org.simple.clinic.storage.DeleteSyncGroupDatabaseAndroidTest
 import org.simple.clinic.storage.PurgeDatabaseAndroidTest
 import org.simple.clinic.storage.migrations.BaseDatabaseMigrationTest
@@ -160,4 +161,5 @@ interface TestAppComponent {
   fun inject(target: QuestionnaireSyncIntegrationTest)
   fun inject(target: QuestionnaireResponseRepositoryAndroidTest)
   fun inject(target: QuestionnaireResponseSyncIntegrationTest)
+  fun inject(target: DatabaseEncryptorTest)
 }