+++ app/src/androidTest/java/org/simple/clinic/user/UserDaoAndroidTest.kt
@@ -9,6 +9,8 @@ import org.junit.Test
 import org.junit.rules.RuleChain
 import org.simple.clinic.AppDatabase
 import org.simple.clinic.TestClinicApp
+import org.simple.clinic.storage.SharedPreferencesMode
+import org.simple.clinic.storage.SharedPreferencesMode.Mode.Default
 import org.simple.sharedTestCode.TestData
 import org.simple.sharedTestCode.util.Rules
 import java.util.UUID
@@ -18,6 +20,7 @@ import javax.inject.Inject
 class UserDaoAndroidTest {
 
   @Inject
+  @SharedPreferencesMode(Default)
   lateinit var sharedPrefs: SharedPreferences
 
   @Inject