+++ app/src/main/java/org/simple/clinic/settings/SettingsInit.kt
@@ -17,6 +17,10 @@ class SettingsInit : Init<SettingsModel, SettingsEffect> {
       effects.add(LoadAppVersionEffect)
     }
 
+    if (model.isDatabaseEncrypted == null) {
+      effects.add(LoadDatabaseEncryptionStatus)
+    }
+
     return first(model, effects)
   }
 }